requirements.txt
async-timeout==4.0.2
python-dotenv==0.21.0
redis==4.4.0
